142750 by sweettea on 2020/03/03 (pair: sclafani) Merge 141734 to Sulfur. 141734 by sweettea on 2019/12/11 (pair: corwin) [VDOSTORY-266] Clarify flush-related physicalLayer functionality.
In times past, PhysicalLayer's isFlushRequired member was used to both
query whether the layer accepted flushes and also whether the VDO was
in sync or async mode. Now that we anticipate adding a third write
policy, we need to query VDO's actual write policy in several places in
the base code.

This change replaces isFlushRequired() with getWritePolicy().

.../physicalLayer.h
Replaced isFlushRequired with getWritePolicy.

.../recoveryJournal.c
.../recoveryJournalBlock.c
.../slabJournal.c
.../vdoPageCache.c
.../vio.c
Updated users of isFlushRequired() to use getWritePolicy() and to have
detailed commentary on why the write policy makes a difference.

Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>