142751 by sweettea on 2020/03/03 (pair: corwin) Merge 141831 to Sulfur. 141831 by sweettea on 2019/12/16 (pair: corwin) [VDOSTORY-266] Add ASYNC_UNSAFE write policy.
This change renames the existing async write policy to async-unsafe,
and adds a new async write policy. The new policy differs only in that
journal blocks aren't written until the data they reference is flushed.
As such, all conditionals on sync/async write policy, except in the
journal, should do the same thing for async-unsafe as async.

Batching journal writes, to reduce the number of flushes, will be
a future performance improvement to the new async mode.

.../types.h
.../compressionState.c
.../recoveryJournal.c
.../vdo.c
.../vioWrite.c
Added new async-unsafe write policy and updated write policy checkers as
appropriate.

.../recoveryJournalBlock.c
Added the meat of the change, which is to make async mode issue a flush
before every journal write to make the data writes stable on disk.

Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>