142785 by sweettea on 2020/03/05 (pair: msakai) Merge 141909 to Sulfur. 141909 by sweettea on 2019/12/19 (pair: corwin) [VDOSTORY-266] Rearrange recovery journal block writing.
This change introduces a WaitQueue for full journal blocks, in close
analogy to the page cache. This enables delaying of full blocks
without having to iterate over all active blocks on each write attempt.
This results in up to 15% better performance in async mode.

.../recoveryJournalInternals.h
.../recoveryJournal.c
Added a WaitQueue for full journal blocks that need writing but aren't
yet writing.
Split writeBlock() into scheduleBlockWrite() and writeBlocks(), to
clarify and consolidate the write policy determinations.
Added batching in async mode -- journal blocks will only be written in
a batch when there are no outstanding block writes.
Rearranged notifyCommitWaiters()/completeWrite() so that pending blocks
are issued only once, at the end of completeWrite(), after possibly
requeueing the just-finished now-full block, to get maximum batching.
Previously, notifyCommitWaiters() might have issued the active block
because it thought there were no pending wakeups, and then the just-
finished partial block might now be full and issue itself for write,
losing potential batching from additional entries in the active
block.

.../recoveryJournalBlock.h
.../recoveryJournalBlock.c
Updated shouldCommit() to be public, as the decision to commit is now
made in recoveryJournal rather than in commitRecoveryBlock(), and
renamed it canCommitRecoveryBlock().
Added Waiter member to allow blocks to go on a WaitQueue.

Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>