142832 by sclafani on 2020/03/10 (pair: msakai) Respelled cast idioms using the kernel's container_of() macro [VDOUP-6]
.../allocatingVIO.h
.../dataVIO.h
Respelled two direct casts that first asserted offsetof() == 0 with
container_of(), but only in places where NULL is implicitly
checked. In both cases, I manually tested with a non-zero offset
to confirm that there are no hidden assumptions elsewhere that
the offset is zero, and also attempted to verify by inspection.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>