142916 by jwiele on 2020/03/16 (pair: msakai) Merge 142550 from //eng/vdo to linux-vdo. 142550 by jwiele on 2020/02/18 (pair: msakai) BZ1787472 VDO-4872 Make vdostats accept short options.
.../vdoStats
Make vdostats accept short (one-character) options. This is
expedited by using argparse instead of the long deprecated
optparse. As a by-product, the usage help message is more
informative and nicer looking.

.../vdostats.8
Document the short options.
Format the man page according to man-pages(7) and matching
other system commands, such as df, and du, insofar as is
practical.

Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>