143039 by jwiele on 2020/03/25 (pair: sweettea) Merge 142946 from linux-uds to Krusty. 142946 by jwiele on 2020/03/18 (pair: msakai) UDSUP-1 UDSUP-6 Simplify the use of thread identifiers.
.../threads.h
Replace ThreadId with pid_t, since both user and kernel space
always use pid_t as the type of a thread identifier; in the
kernel, it's the pid of the current task, in user space it's
the pid_t returned by the gettid syscall.
Rename getThreadId to get_thread_id to match kernel style.

.../loggerLinuxUser.c
.../minisyslog.c
.../threadsLinuxUser.c
Replace ThreadId with pid_t and getThreadId with get_thread_id.

143049 by jwiele on 2020/03/25 (pair: sclafani)
Fix thread ID to match krusty change 143039.

Change 143039 is the merge of 142946 from linux-uds to krusty:
UDSUP-1 UDSUP-6 Simplify the use of thread identifiers.

.../threads.h
Rename getThreadId to get_thread_id to match UDS.

.../trace.c
Accomodate the rename of getThreadId to get_thread_id.

Co-authored-by: jwiele <jwiele@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>