143394 by corwin on 2020/04/15 (pair: sweettea) Eliminate AsyncEnqueueable [VDOUP-3].
The first real step in eliminating enqueueables is to modify AsyncLayer
to not use them. This change does so by embedding a kvdo_work_item in
completion, and defining a version of that structure for use in user
space.

.../completion.h
Added a definition of struct kvdo_work_item which, for now, simply
contains a single FunnelQueueEntry. This definition applies when
compiled with -DUSER_SPACE. Otherwise, workQueue.h is included to
provide the appropriate definition.
Added a 'work_item' field to completion which is a kvdo_work_item.

.../physicalLayer.h
Modified enqueuer() to take a completion instead of an enqueueable.

.../completion.c
Reflected change to enqueuer(). The assertion used to catch attempts to
enqueue completions without enqueueables has moved into the kernel
layer's implementation of the enqueuer() method.

.../adminCompletion.c
Reflected change to enqueuer().

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>