143596 by corwin on 2020/04/23 (pair: sweettea) Remove the linear block map (except from the block map state).
Since we never released a VDO with a hybrid block map, and we have no
plans to upgrade sodium or older VDOs to magnesium and beyond, the
linear portion of the block map no longer exists. This change
eliminates all references to it except in the block map state itself,
since that would entail an on-disk format change.

.../blockMap.h
.../blockMapInternals.h
.../blockMap.c
Removed 'flat_page_count' constructor parameter and field.

.../blockMap.c
.../blockMapTree.c
.../forest.c
.../vdoConfig.c
Reflected removal of the field and constructor parameter.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>