143640 by sclafani on 2020/04/24 (pair: sweettea) Converted one third of vdo/base to use __must_check [VDOUP-7]
This is a script-driven substitution of the kernel's __must_check
macro for all uses in vdo/base/[a-m]* of
__attribute__((warn_unused_result)). The majority practice in the
kernel is to place it immediately in front of the function name,
instead of first or last as we have always done, so also
re-arranged declarations to conform. See UDS change 143379 for
more details.

.../actionManager.h
.../adminCompletion.h
.../adminState.c
.../adminState.h
.../allocationSelector.h
.../blockAllocator.c
.../blockAllocator.h
.../blockAllocatorInternals.h
.../blockMap.c
.../blockMap.h
.../blockMapInternals.h
.../blockMapPage.h
.../blockMapRecovery.c
.../blockMapTree.c
.../blockMapTree.h
.../blockMapTreeInternals.h
.../completion.c
.../completion.h
.../compressionState.c
.../compressionState.h
.../dataVIO.h
.../dirtyLists.h
.../dirtyListsInternals.h
.../extent.h
.../fixedLayout.h
.../flush.h
.../forest.h
.../hashLock.h
.../hashLockInternals.h
.../hashZone.h
.../header.h
.../intMap.h
.../lockCounter.h
.../logicalZone.h
Replaced __attribute__((warn_unused_result)) with __must_check.
Moved __must_check to where the kernel usually puts it.
Re-filled modified function declaration, stripping alignment
whitespace and re-indenting parameters using tabs.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>