143645 by sclafani on 2020/04/24 (pair: sweettea) Converted another third of vdo/base to use __must_check [VDOUP-7]
This is a script-driven substitution of the kernel's __must_check
macro for all uses in vdo/base/[n-r]* of
__attribute__((warn_unused_result)). The majority practice in the
kernel is to place it immediately in front of the function name,
instead of first or last as we have always done, so also
re-arranged declarations to conform. See UDS change 143379 for
more details.

.../numUtils.h
.../packedRecoveryJournalBlock.h
.../packer.c
.../packer.h
.../partitionCopy.c
.../partitionCopy.h
.../pbnLock.h
.../pbnLockPool.h
.../physicalZone.h
.../pointerMap.h
.../priorityTable.h
.../readOnlyNotifier.h
.../readOnlyRebuild.c
.../recoveryJournal.c
.../recoveryJournal.h
.../recoveryJournalBlock.c
.../recoveryJournalBlock.h
.../recoveryJournalInternals.h
.../recoveryUtils.c
.../recoveryUtils.h
.../refCounts.c
.../refCounts.h
.../refCountsInternals.h
.../referenceCountRebuild.c
.../referenceOperation.h
.../slab.h
.../slabDepot.c
.../slabDepot.h
.../slabDepotInternals.h
.../slabJournal.c
.../slabJournal.h
.../slabJournalInternals.h
.../slabScrubber.c
.../slabScrubber.h
.../slabSummary.c
.../slabSummary.h
.../superBlock.c
.../superBlock.h
Replaced __attribute__((warn_unused_result)) with __must_check.
Moved __must_check to where the kernel usually puts it.
Re-filled modified function declaration, stripping alignment
whitespace and re-indenting parameters using tabs.

Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>